Created a Python GUI Launcher for the UI
run-launcher.py is the script for creating the GUI and launching the commands to launch the web UI.
start-webui.bat launches to .py as a one click launch

This code defines a GUI application that enables users to select a machine learning model and configure several options for running the web ui server. The GUI contains multiple listboxes and entry fields that allow users to specify the port number, UI mode, and extensions to be used by the server. When the user clicks the "Run" button, the application launches a command prompt window and runs a Python script that starts the server using the specified options. Additionally, the application opens a web browser window that displays the server's user interface. Finally, the application provides a "Shutdown" button that terminates the running server and closes the associated command prompt window.

Video of the app here: https://streamable.com/7gjizs